refactor: 情報 API Router へのモジュール化
